fix(workflow): Download TF-generated JS and index.html before syncing to S3 to prevent deletion
The s3_sync's `delete` option would delete the TF-generated view-count.js file as it is not in our repo. To prevent this, we must download it from the S3 bucket before syncing. This ensures that the file is present during the sync and is not mistakenly deleted.

Added `set +e` before the `grep` command so that the workflow won't stop if `grep` failed (no match found).